Contacts Challenge
Initially used the Linq library to try and pass the Hackerrank challenge for Contacts.

Unfortunately  this method had very adverse performance issues and several cases threw timeout exceptions.

I attempted to use a Trie to store the substrings while also incrementing the occurrences of these substrings.

A simpler solution was to use a dictionary to store all values in a loop, and then do a simple look up based on them.